Initial guide for updating sources
Note that I had to use `index.md` instead of `README.md` to refer to the main page for "Updating Sources". Seems like a bug in mdBook.

Also, I'm not sure if adding HTML anchors to the headers in "Updating Sources"'s README is a good idea, but it's better than copying the entire step summary as a link and pasting that everywhere I need to refer to the general steps in the additional guidelines for each project. For now, it's not a big deal because only `libc` is there, but there's a bunch more projects that need additional guidelines, so it's not ideal to copy-paste the long header names as links in all of them.